Make debounce function a utility function and minor bug fixes
